Folder structure added to project
