+++ app.py
@@ -1,334 +1,21 @@
 from flask import Flask
-from flask_restful import Api, Resource, reqparse, abort
+from flask_restful import Api
 
 # Importamos los archivos que necesitamos
 import sys
 sys.path.append(".")
-from graph import Graph
-from priority_queue import PriorityQueue
+#from graph import Graph
+from Graph.graph_api import GraphApi
+from PriorityQueue.priority_queue_api import PriorityQueueApi
+#from priority_queue import PriorityQueue
 
 # Creamos la aplicación
 app = Flask(__name__)
 api = Api(app)
 
-# Configuramos los parámetros del necesarios para crear un grafo
-graph_put_args = reqparse.RequestParser()
-graph_put_args.add_argument("start", type=int, help="Start node is required", required=True)
-graph_put_args.add_argument("goal", type=int, help="Goal node is required", required=True)
-graph_put_args.add_argument("edges", type=str, help="Edges are required", required=True)
 
-temp_graph = []
-
-
-class GraphApi(Resource):
-    def __init__(self):
-        # Arcos
-        self._arcs = []
-
-        # Nodo inicial
-        self._start = None
-
-        # Nodo final
-        self._goal = None
-
-        # Instancia de la clase 'Graph'
-        self._g_instance = None
-
-        # Mensaje de respuesta al usuario
-        self._data = None
-
-
-    # Getters y setters
-    @property
-    def arcs(self):
-        '''
-        Getter de la variable 'arcs'.
-        :return: La variable 'arcs'
-        '''
-        return self._arcs
-
-    @arcs.setter
-    def arcs(self, arcs):
-        '''
-            Setter de la variable 'arcs'.
-            :return: None
-        '''
-        self._arcs = arcs
-
-    @property
-    def start(self):
-        '''
-            Getter de la variable 'start'.
-            :return: La variable 'start'
-        '''
-        return self._start
-
-    @start.setter
-    def start(self, start):
-        '''
-            Setter de la variable 'start'.
-            :return: None
-        '''
-        self._start = start
-
-    @property
-    def goal(self):
-        '''
-            Getter de la variable 'goal'.
-            :return: La variable 'goal'
-        '''
-        return self._goal
-
-    @goal.setter
-    def goal(self, goal):
-        '''
-            Setter de la variable 'goal'.
-            :return: None
-        '''
-        self._goal = goal
-
-    @property
-    def g_instance(self):
-        '''
-            Getter de la variable 'g_instance'.
-            :return: La variable 'g_instance'
-        '''
-        return self._g_instance
-
-    @g_instance.setter
-    def g_instance(self, g_instance):
-        '''
-            Setter de la variable 'g_instance'.
-            :return: None
-        '''
-        self._g_instance = g_instance
-
-    @property
-    def data(self):
-        '''
-            Getter de la variable 'data'.
-            :return: La variable 'data'
-        '''
-        return self._data
-
-    @data.setter
-    def data(self, data):
-        '''
-            Setter de la variable 'data'.
-            :return: None
-        '''
-        self._data = data
-
-    def get(self):
-        '''
-        Función que devuelve la lista de arcos del árbol (si la hay) al hacer
-        una petición GET al endpoint '/graph'.
-        :return: Un JSON con los arcos del grafo.
-        '''
-        print("Self.arcs: {}".format(self.arcs))
-        # Verificamos que los arcos no estén vacíos
-        self.are_arcs_empty()
-
-        return {"data": temp_graph}
-
-    def put(self):
-        '''
-        Función encargada de recibir los datos necesarios para crear un grafo.
-        Los recibe en formato JSON cuando se hace una petición PUT al endpoint.
-        :return: Una tupla de la forma (num_personas_necesarias, [lista_de_personas_necesarias])
-        '''
-        # Verificamos que cumpla con los datos que necesitamos
-        args = graph_put_args.parse_args()
-
-        # Función encargada de tratar la entrada
-        self.clean_data(args)
-
-        # Iniciamos el grafo
-        self.start_graph()
-
-        # Le damos formato al JSON de respuesta
-        self.data = {
-            "distance": self.data[0],
-            "nodes": self.data[1]
-        }
-
-        return self.data, 201
-
-    def clean_data(self, args):
-        # Guardamos el inicio y el objetivo
-        self.start = args['start']
-        self.goal = args['goal']
-
-        # Separamos los arcos por parejas
-        edges = args['edges'].split(',')
-
-        # Separamos cada pareja
-        edges = [tuple(map(int, edge.split('-'))) for edge in edges]
-
-        # Guardamos los arcos en su variable respectiva
-        self.arcs = edges
-
-        # Modificamos la variable global
-        for el in self.arcs:
-            temp_graph.append(el)
-
-    def delete(self):
-        temp_graph.clear()
-        return {"message": "Arcos borrados satisfactoriamente"}
-
-    def are_arcs_empty(self):
-        if len(temp_graph) == 0:
-            abort(404, message="No se puede consultar un grafo sin nodos")
-
-    def start_graph(self):
-        self.g_instance = Graph(self.goal + 1)
-
-        # Agregamos los arcos al grafo
-        for pair in self.arcs:
-            self.g_instance.add_edge(pair[0], pair[1])
-
-        # Buscamos la cantidad de personas necesarias para llegar de 'start' a 'goal'
-        self.data = self.g_instance.find_love(self.start, self.goal)
-
-
-# Generamos el endpoint
+# Generamos el endpoint para la funcionalidad de grafos
 api.add_resource(GraphApi, "/graph")
 
-
-# Configuramos los parámetros necesarios para crear una cola de prioridad
-p_queue_args = reqparse.RequestParser()
-p_queue_args.add_argument("names", type=str, help="The name of the person is required", required=True)
-p_queue_args.add_argument("popularity", type=str, help="The popularity of the person is required", required=True)
-p_queue_args.add_argument("times_spoken", type=str, help="The number of times this person have spoken with the goal is "
-                                                         "required", required=True)
-
-temp_priority_queue = []
-
-
-class PriorityQueueApi(Resource):
-    def __init__(self):
-        # Instancia de la clase 'PriorityQueue'
-        self._pq = None
-
-        # Lista con las parejas
-        self._data = []
-
-    # Setters y getters
-    @property
-    def pq(self):
-        '''
-            Getter de la variable 'pq'
-            :return: Variable 'pq'
-        '''
-        return self._pq
-
-    @pq.setter
-    def pq(self, pq):
-        '''
-        Setter de la variable 'pq'
-        :param pq: Instancia de la clase "PriorityQueue"
-        :return: None
-        '''
-        self._pq = pq
-
-    @property
-    def data(self):
-        '''
-            Getter de la variable 'data'
-            :return: Variable 'data'
-        '''
-        return self._data
-
-    @data.setter
-    def data(self, data):
-        '''
-        Setter de la variable 'data'
-        :param pq: Instancia de la clase "PriorityQueue"
-        :return: None
-        '''
-        self._data = data
-
-    def put(self):
-        '''
-        Función encargada de recibir los datos del Frontend.
-        :return:
-        '''
-        # Verificamos que cumpla con los datos que necesitamos
-        args = p_queue_args.parse_args()
-
-        # Limpiamos los datos
-        self.clean_data(args)
-
-        # Creamos la lista de prioridad
-        self.data = self.create_priority_queue(self.data)
-
-        # Le damos formato al JSON de respuesta
-        self.data = {
-            "name": self.data[0][0],
-            "score": self.data[0][1]
-        }
-
-        # Devolvemos la lista adecuada y el estado de éxito
-        return self.data, 201
-
-    def clean_data(self, args):
-        '''
-        Función encargada de tomar el JSON con los datos recibidos a través de la petición PUT
-        al endpoint '/priority_queue' y devolver una lista de tuplas de la forma: (nombre, factor_de_importancia)
-        :param args: Diccionario que contiene el JSON con los nombres, índices de popularidad y número de contactos con la
-        persona objetivo en la última semana.
-        :return: Lista de tuplas.
-        '''
-
-        # Tratamos los datos para poder agruparlos
-        names = args['names'].split(",")
-
-        popularity = list(map(int, args['popularity'].split(',')))
-
-        times_spoken = list(map(int, args["times_spoken"].split(',')))
-
-        # Verificamos que las listas tengan el mismo tamaño
-        self.are_equal_length(names, popularity, times_spoken)
-
-        # Creamos la lista de tuplas y calculamos el factor de importancia
-        # Factor de importancia =(índice de importancia x número de contactos en la última semana)
-        temp = []
-
-        for i in range(len(names)):
-            temp_list = [None, None]
-            temp_list[0] = names[i]
-            temp_list[1] = popularity[i]*times_spoken[i]
-
-            temp.append(temp_list)
-
-        # Guardamos la lista en la variable respectiva
-        self.data = temp
-
-    def are_equal_length(self, names, popularity, times_spoken):
-        '''
-        Esta función aborta del proceso en caso de no tener la misma cantidad de datos en las tres
-        listas ingresadas como parámetro.
-        :param names: Lista de nomrbes.
-        :param popularity: Lista de índices de popularidad.
-        :param times_spoken: Lista de contactos en la última semana.
-        :return: None
-        '''
-
-        if len(names) != len(popularity) or len(names) != len(times_spoken) or len(popularity) != len(times_spoken):
-            abort(411, message="Las listas no tienen el mismo tamaño, verifique los datos ingresados")
-
-    def create_priority_queue(self, data):
-        '''
-        Esa función crea la cola de prioridad y devuelve el top 5 (si los har) de personas importantes,
-        en caso de tener menos de 5 datos, devuelve los datos ingresados en orden decreciente.
-        :param data: Lista donde cada elemento es una tupla de la forma (nombre, importancia).
-        :return: Una lista organizada de acuerdo al factor de importancia (de mayor a menor).
-        '''
-
-        self.pq = PriorityQueue(data, 1)
-
-        return self.pq.get_top()
-
-
-# Generamos el endpoint
+# Generamos el endpoint para la funcionalidad de cola de prioridad
 api.add_resource(PriorityQueueApi, "/priority_queue")
